+++ src/models/ticker.rs
@@ -1,9 +1,10 @@
 //! Ticker (price summary) channel models.
 
+use rust_decimal::Decimal;
 use serde::Deserialize;
 
 /// An update message from the `ticker` channel.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct TickerUpdateResponse {
     pub channel: String,
     #[serde(rename = "type")]
@@ -12,23 +13,23 @@ pub struct TickerUpdateResponse {
 }
 
 /// Real-time ticker snapshot for a single trading pair.
-#[derive(Deserialize)]
+#[derive(Debug, Clone, Deserialize)]
 pub struct TickerData {
     pub symbol: String,
-    pub bid: f64,
+    pub bid: Decimal,
     /// Quantity available at the best bid price.
-    pub bid_qty: f64,
-    pub ask: f64,
+    pub bid_qty: Decimal,
+    pub ask: Decimal,
     /// Quantity available at the best ask price.
-    pub ask_qty: f64,
-    pub last: f64,
-    pub volume: f64,
+    pub ask_qty: Decimal,
+    pub last: Decimal,
+    pub volume: Decimal,
     /// Volume-weighted average price.
-    pub vwap: f64,
-    pub low: f64,
-    pub high: f64,
+    pub vwap: Decimal,
+    pub low: Decimal,
+    pub high: Decimal,
     /// Absolute price change over the last 24 hours.
-    pub change: f64,
+    pub change: Decimal,
     /// Price change as a percentage over the last 24 hours.
-    pub change_pct: f64,
+    pub change_pct: Decimal,
 }
